+++ utils/i18n/fr/translation.ts
@@ -59,6 +59,7 @@ const translation = {
     famousart: "Œuvres d’art parmi les plus célèbres...",
     name: "Nom(s) :",
     names: "Prénom(s), nom(s) :",
+    talentsart: "Quelques œuvres d’artistes talentueux...",
   },
   slider: {
     value_large: "An ",